Lots of work on the seamless mode. Now starts to be really useful.
It can detect more types of events, and has visual clues over if the desired event will be properly recorded or not.
It shows in green elements that can be uniquely recognized, and red the not "selectable" tags.
Also added initial version of autocompletion that helps with commands and selectors.
Ctest functions are now called through commands, making more dificult to distinguish them, allowing for example to replace a command with a function (load, for example).